Separated main and network bundles
