+++ assets/src/scripts/components/hydrograph/date-controls.spec.js
@@ -1,6 +1,6 @@
 import { select } from 'd3-selection';
 
-import { Actions, configureStore } from '../../store';
+import { Actions, configureStore } from '../../store/site-store';
 
 import {drawDateRangeControls} from './date-controls';
