+++ assets/src/scripts/components/hydrograph/graph-controls.spec.js
@@ -1,4 +1,4 @@
-import {Actions, configureStore} from '../../store';
+import {Actions, configureStore} from '../../store/site-store';
 import { select } from 'd3-selection';
 import { drawGraphControls } from './graph-controls';
