+++ assets/src/scripts/components/map/flood-slider.spec.js
@@ -1,5 +1,5 @@
 import { select } from 'd3-selection';
-import { configureStore } from '../../store';
+import { configureStore } from '../../store/site-store';
 import { floodSlider } from './flood-slider';
