+++ assets/src/scripts/components/map/index.js
@@ -6,7 +6,7 @@ import { link } from '../../lib/d3-redux';
 import config from '../../config';
 import { FLOOD_EXTENTS_ENDPOINT, FLOOD_BREACH_ENDPOINT, FLOOD_LEVEE_ENDPOINT } from '../../web-services/flood-data';
 import { hasFloodData, getFloodExtent, getFloodStageHeight } from '../../selectors/flood-data-selector';
-import { Actions } from '../../store';
+import { Actions } from '../../store/site-store';
 import { floodSlider } from './flood-slider';
 import { createLegendControl, createFIMLegend, createNldiLegend } from './legend';
 import { addNldiLayers} from './nldiMapping';
