+++ assets/src/scripts/networks/components/network-mapping.js
