+++ assets/src/scripts/networks/network-data-reducer.js
