+++ assets/src/scripts/networks/network-data-reducer.spec.js
