+++ assets/src/scripts/networks/network-data-selector.js
