+++ assets/src/scripts/networks/network-data-selector.spec.js
