+++ assets/src/scripts/networks/network-data.js
