+++ assets/src/scripts/networks/network-data.spec.js
