+++ assets/src/styles/network.scss
@@ -1,0 +1,147 @@
+@import 'leaflet';
+@import 'common';
+
+$fa-font-path: './fonts' !default;
+@import '../../node_modules/@fortawesome/fontawesome-free/scss/fontawesome';
+@import '../../node_modules/@fortawesome/fontawesome-free/scss/brands';
+@import '../../node_modules/@fortawesome/fontawesome-free/scss/solid';
+
+@import 'partials/script';
+@import 'partials/slider';
+@import 'partials/social-share';
+@import 'partials/tooltip';
+
+@import '../../node_modules/markercluster-iow/dist/MarkerCluster';
+@import '../../node_modules/markercluster-iow/dist/MarkerCluster.Default';
+
+.wdfn-component {
+  &[data-component='network-map'] {
+    #network-map {
+      height: 350px;
+      margin: 1em 0;
+
+      .legend {
+        padding: 2px;
+        background-color: white;
+
+        .legend-expand-container {
+          font-size: 14px;
+          font-weight: bold;
+
+          span {
+            padding-right: 2px;
+          }
+
+          .legend-expand {
+            width: auto;
+            float: right;
+            padding: 0;
+            margin: 0;
+            color: black;
+            border: none;
+            border-radius: 0;
+            box-shadow: none;
+
+            i {
+              padding: 2px;
+              background-color: color('base-lighter');
+            }
+          }
+        }
+
+        .legend-list-container {
+          #network-legend-list {
+            span {
+              margin-left: 2px;
+            }
+
+            img {
+              padding-top: 2px;
+              width: 15px;
+            }
+          }
+
+          #fim-legend-list {
+            span {
+              margin-left: 2px;
+            }
+
+            img {
+              padding-top: 2px;
+              width: 15px;
+              height: 13px;
+            }
+          }
+        }
+
+      }
+
+      .leaflet-control-layers-selector {
+        left: 0;
+        position: relative;
+        display: inline;
+        height: auto;
+        width: auto;
+        outline-style: none;
+      }
+
+      label {
+        margin: auto;
+      }
+    }
+    margin-bottom: 1em;
+  }
+}
+
+/* This fixes an issue with Firefox not hiding long divs. I opened https://github.com/uswds/uswds/issues/2380 .
+ * If it gets addressed this can be removed.
+ */
+.usa-sr-only {
+  overflow: hidden;
+}
+
+body {
+  #main-content {
+    margin-top: 1em;
+    margin-bottom: 1em;
+    padding: 0 1em;
+  }
+
+  .content-container {
+    margin-left: auto;
+    margin-right: auto;
+    @include u-maxw('desktop');
+  }
+
+  #wdfn-alert-banner input:checked ~ * {
+    display: none;
+  }
+
+  #wdfn-alert-banner {
+    background-color: color('accent-cool-lighter');
+
+    .wdfn-alert-body {
+      padding-top: 0.5em;
+      padding-bottom: 0.5em;
+      padding-right: 1em;
+      padding-left: 1em;
+      margin-top: 0;
+
+      .close {
+        float: right;
+        cursor: pointer;
+        margin-top: -5px;
+        margin-right: -10px;
+      }
+
+      .div {
+        font-size: 0.9em;
+        @include at-media('tablet') {
+          font-size: 1em;
+        }
+        padding-left: 1em;
+      }
+    }
+
+  }
+}
