--- assets/src/styles/networks/_networkmap.scss
@@ -1,73 +1,0 @@
-#network-map {
-  height: 350px;
-  margin: 1em 0;
-
-  .legend {
-    padding: 2px;
-    background-color: white;
-
-    .legend-expand-container {
-      font-size: 14px;
-      font-weight: bold;
-
-      span {
-        padding-right: 2px;
-      }
-
-      .legend-expand {
-        width: auto;
-        float: right;
-        padding: 0;
-        margin: 0;
-        color: black;
-        border: none;
-        border-radius: 0;
-        box-shadow: none;
-
-        i {
-          padding: 2px;
-          background-color: color('base-lighter');
-        }
-      }
-    }
-
-    .legend-list-container {
-      #network-legend-list {
-        span {
-          margin-left: 2px;
-        }
-
-        img {
-          padding-top: 2px;
-          width: 15px;
-        }
-      }
-
-      #fim-legend-list {
-        span {
-          margin-left: 2px;
-        }
-
-        img {
-          padding-top: 2px;
-          width: 15px;
-          height: 13px;
-        }
-      }
-    }
-
-  }
-
-  .leaflet-control-layers-selector {
-    left: 0;
-    position: relative;
-    display: inline;
-    height: auto;
-    width: auto;
-    outline-style: none;
-  }
-
-  label {
-    margin: auto;
-  }
-}
