+++ wdfn-server/waterdata/templates/base_network.html
@@ -1,0 +1,21 @@
+{% extends 'base_plain_network.html' %}
+
+{% block body %}
+    <section role="banner">
+        {% include 'partials/header.html' %}
+        <header id="wdfn-alert-banner">
+            <input type="checkbox" id="wdfn-alert-dismiss" hidden/>
+            <div class="wdfn-alert-body content-container">
+                <label title="Close" class="close" for="wdfn-alert-dismiss"><i class="fas fa-times"></i></label>
+                <div>
+                    Beta release. Use <a class="usa-link" target="_blank" href="https://water.usgs.gov/contact/gsanswers?pemail=gs-w_water_data_for_the_nation&subject=Water%20Data%20for%20the%20Nation%20Feedback&viewnote=%3CH1%3EUSGS+WDFN+TNG+Feedback%3C/H1%3E">this form</a> to provide feedback.
+                </div>
+            </div>
+        </header>
+    </section>
+    <main id="main-content" class="grid-container content-container usa-prose">
+        {% block content %}{% endblock %}
+    </main>
+    {% include 'partials/footer.html' %}
+    {% block extra_body %}{% endblock %}
+{% endblock %}
