+++ wdfn-server/waterdata/templates/base_plain_network.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <meta name="format-detection" content="telephone=no">
+        {% block extra_head_tags %}{% endblock %}
+
+        {% include 'partials/google_analytics.html' %}
+
+        <link rel="shortcut icon" type="image/ico" href="{{ 'img/usgs_favicon.ico' | asset_url }}">
+        <link rel="stylesheet" href="{{ 'network.css' | asset_url }}">
+        <script type="text/javascript">
+            var CONFIG = {
+                'SERVICE_ROOT': '{{ config.SERVICE_ROOT }}/nwis',
+                'PAST_SERVICE_ROOT': '{{ config.PAST_SERVICE_ROOT }}/nwis',
+                'OBSERVATIONS_ENDPOINT': '{{ config.OBSERVATIONS_ENDPOINT }}',
+                'HYDRO_ENDPOINT': '{{ config.HYDRO_ENDPOINT }}',
+                'FIM_GIS_ENDPOINT': '{{ config.FIM_GIS_ENDPOINT }}',
+                'FIM_ENDPOINT': '{{ config.FIM_ENDPOINT }}',
+                'STATIC_URL': '{{ config.STATIC_ROOT }}',
+                'TIMESERIES_AUDIO_ENABLED': '{{ config.TIMESERIES_AUDIO_ENABLED}}' === 'True' ? true : false,
+                'MULTIPLE_TIME_SERIES_METADATA_SELECTOR_ENABLED': '{{ config.MULTIPLE_TIME_SERIES_METADATA_SELECTOR_ENABLED }}' === 'True' ? true : false,
+                'WEATHER_SERVICE_ROOT': '{{ config.WEATHER_SERVICE_ROOT }}',
+                'CITIES_ENDPOINT': '{{ config.CITIES_ENDPOINT }}',
+                'NLDI_SERVICES_ENDPOINT': '{{ config.NLDI_SERVICES_ENDPOINT }}',
+                'NLDI_SERVICES_DISTANCE': '{{ config.NLDI_SERVICES_DISTANCE }}',
+                'NETWORK_ENDPOINT': '{{ config.NETWORK_ENDPOINT }}'
+            };
+        </script>
+        <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
+        <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.en"></script>
+        <script type="text/javascript">
+            if (!!document.documentMode) {
+                document.write('<script src="{{ "scripts/date-time-format-timezone-complete-min.js" | asset_url }}"><\/script>');
+            }
+        </script>
+
+
+        <script async src="{{ 'network-bundle.js' | asset_url }}"></script>
+        {% block page_script %}{% endblock %}
+        <!-- Google Tag Manager --><script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src='https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);})(window,document,'script','dataLayer','GTM-TKQR8KP');</script>
+        <title>
+            {% block title %}{% if page_title %}{{ page_title }} - {% endif %}USGS Water Data for the Nation{% endblock %}
+        </title>
+    </head>
+    <body {% if body_id %}id="{{ body_id }}"{% endif %}>
+        <script>document.body.className += ' js';</script>
+        {% block body %}{% endblock %}
+    </body>
+    {% if config.LIVE_RELOAD_PATH %}
+        <script type="text/javascript">
+            // <![CDATA[  <-- For SVG support
+            if ('WebSocket' in window) {
+                (function() {
+                    function refreshCSS() {
+                        var sheets = [].slice.call(document.getElementsByTagName("link"));
+                        var head = document.getElementsByTagName("head")[0];
+                        for (var i = 0; i < sheets.length; ++i) {
+                            var elem = sheets[i];
+                            head.removeChild(elem);
+                            var rel = elem.rel;
+                            if (elem.href && typeof rel != "string" || rel.length == 0 || rel.toLowerCase() == "stylesheet") {
+                                var url = elem.href.replace(/(&|\?)_cacheOverride=\d+/, '');
+                                elem.href = url + (url.indexOf('?') >= 0 ? '&' : '?') + '_cacheOverride=' + (new Date().valueOf());
+                            }
+                            head.appendChild(elem);
+                        }
+                    }
+                    var socket = new WebSocket('{{ config.LIVE_RELOAD_PATH }}');
+                    socket.onmessage = function(msg) {
+                        if (msg.data == 'reload') window.location.reload();
+                        else if (msg.data == 'refreshcss') refreshCSS();
+                    };
+                    console.log('Live reload enabled.');
+                })();
+            }
+            // ]]>
+        </script>
+    {% endif %}
+    {% block extra_html %}{% endblock %}
+    <!-- Google Tag Manager (noscript) --><noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-TKQR8KP" height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
+</html>
