+++ wdfn-server/waterdata/templates/networks.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'base_network.html' %}
 
 {% if network_cd %}{% set page_title = 'Networks' %}{% endif %}
